code refactor

+++ chat/repository/src/main/java/com/devfalah/repository/mappers/NotificationMapper.kt
@@ -21,7 +21,7 @@ fun NotificationDataModel.toMessage(): Message {
         id = this.id,
         friendId = this.friendId,
         fromMe = false,
-        time = this.time,
+        time = this.time.toLong(),
         message = this.messageText
     )
 }
